Add preprocessing for metadata calls
At present, Presto analyzer makes sequential calls to fetch metadata
for all the tables/views used in the query. The total time spent in
the metadata calls is equal to time taken by each call.

This can be overall very expansive, depending on the total number of
metadata calls and tail latencies for the metadata calls.

In this change, we are adding a feature to preprocess required metadata
information required for the analysis. This feature is controlled by
session property `pre_process_metadata_calls`.
The JMX metrics for executor used for preprocessing can be tracked at
`presto.com.facebook.presto.sql.analyzer:name=MetadataExtractor:Executor.PoolSize`

DDL statement paths are not covered int this flow at the moment.