[Presto-on-Spark] Move DetachedNativeExecutionProcess from test to src
Currently DetachedNativeExecutionProcess class is defined in test folder
as it was only being used for test execution. But now that we want to
support Presto-on-Spark in localMode, we want to be able to talk
to an already running CPP process during e2e query execution. Moving it
out into src to enable this